+++ appcheck/firebase-appcheck-debug/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Unreleased
+
+# 16.1.2
+* [unchanged] Updated to keep [app_check] SDK versions aligned.
+
+# 16.1.1
 * [changed] Migrated [app_check] SDKs to use standard Firebase executors. (#4431, #4449)
 * [changed] Integrated the [app_check] Debug SDK with Firebase Components. (#4436)
 * [changed] Moved Task continuations off the main thread. (#4453)